Skin_Cancer_Classification_Web_Using_DeepLearning-master/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.2.0
albumentations>=1.1.0
scikit-learn>=0.24.0
tensorflow>=2.5.0
flask>=2.0.0